don't gen anonymous classes when reporting manual errors
we were seeing an uptick in OOMs in r101-api around the same time we
released https://github.com/remind101/exceptions/pull/13. after
reverting that, the OOMs went away. so it could very well be the
anonymous classes causing the memory issues.

i figured out a way to do this without creating anonymous classes. it's
just as hax as the last approach, if not more hax, but the rollbar gem
is happy with it.

![proof](http://s.goobtown.net/ag9zfmdvb2JzaG90cy1ocmRyDQsSBFNob3QYkomIBQw)